+++ android/testng-examples/pom.xml
@@ -102,7 +102,7 @@
                         <artifactId>maven-surefire-plugin</artifactId>
                         <configuration>
                             <suiteXmlFiles>
-                                <suiteXmlFile>src/test/resources/com/browserstack/run_first_test/first.testng.xml</suiteXmlFile>
+                                <suiteXmlFile>src/test/resources/com/browserstack/first.testng.xml</suiteXmlFile>
                             </suiteXmlFiles>
                         </configuration>
                     </plugin>
@@ -118,7 +118,7 @@
                         <artifactId>maven-surefire-plugin</artifactId>
                         <configuration>
                             <suiteXmlFiles>
-                                <suiteXmlFile>src/test/resources/com/browserstack/run_first_test/local.testng.xml</suiteXmlFile>
+                                <suiteXmlFile>src/test/resources/com/browserstack/local.testng.xml</suiteXmlFile>
                             </suiteXmlFiles>
                         </configuration>
                     </plugin>
