+++ android/testng-examples/src/test/resources/com/browserstack/first.testng.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
-<suite name="Local">
-    <test name="LocalTest">
+<suite name="First">
+    <test name="FirstTest">
         <classes>
-            <class name="com.browserstack.run_first_test.LocalTest"/>
+            <class name="com.browserstack.FirstTest"/>
         </classes>
     </test>
 </suite>
