+++ ios/testng-examples/src/test/resources/com/browserstack/local.testng.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
-<suite name="First">
-    <test name="FirstTest">
+<suite name="Local">
+    <test name="LocalTest">
         <classes>
-            <class name="com.browserstack.run_first_test.FirstTest"/>
+            <class name="com.browserstack.LocalTest"/>
         </classes>
     </test>
 </suite>
